Login, melhorias de tratamento e relacionamento

+++ ProjetoCafe/projeto-cafe/routes/api.php
@@ -2,6 +2,7 @@
 
 use App\Http\Controllers\CompraController;
 use App\Http\Controllers\FilaController;
+use App\Http\Controllers\LoginController;
 use App\Http\Controllers\UsuarioController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
@@ -10,32 +11,36 @@ Route::prefix('usuario') -> group(function (){
     Route::get('', [UsuarioController::class, 'listar']);
     Route::get('/{id}', [UsuarioController::class, 'buscarId']);
     Route::post('/criar', [UsuarioController::class, 'criar']);
-    Route::put('/atualizar/{id}', [UsuarioController::class, 'atualizar']);
-    Route::delete('/deletar/{id}', [UsuarioController::class, 'deletar']);
-    Route::delete('/destroy/{id}', [UsuarioController::class, 'destroy']);
-    Route::post('/restore/{id}', [UsuarioController::class, 'restore']);
+    Route::middleware('auth:sanctum')->put('/atualizar/{id}', [UsuarioController::class, 'atualizar']);
+    Route::middleware('auth:sanctum')->delete('/deletar/{id}', [UsuarioController::class, 'deletar']);
+    Route::middleware(['auth:sanctum', 'admin'])->delete('/destroy/{id}', [UsuarioController::class, 'destroy']);
+    Route::middleware(['auth:sanctum', 'admin'])->post('/restore/{id}', [UsuarioController::class, 'restore']);
 });
 
 Route::prefix('fila') -> group(function (){
     Route::get('', [FilaController::class, 'listar']);
     Route::get('/{id}', [FilaController::class, 'buscarId']);
-    Route::post('/criar/{id}', [FilaController::class, 'criar']);
-    Route::delete('/deletar/{id}', [FilaController::class, 'deletar']);
-    Route::delete('/destroy/{id}', [FilaController::class, 'destroy']);
-    Route::post('/restore/{id}', [FilaController::class, 'restore']);
-    Route::post('/moveraposcompra/{id}', [FilaController::class, 'moverAposCompra']);
+    Route::middleware('auth:sanctum')->post('/criar', [FilaController::class, 'criar']);
+    Route::middleware('auth:sanctum')->delete('/deletar/{id}', [FilaController::class, 'deletar']);
+    Route::middleware(['auth:sanctum', 'admin'])->delete('/destroy/{id}', [FilaController::class, 'destroy']);
+    Route::middleware(['auth:sanctum', 'admin'])->post('/restore/{id}', [FilaController::class, 'restore']);
+    Route::get('/proximo', [FilaController::class, 'proximo']);
+    Route::get('/posicao/{usuario_id}', [FilaController::class, 'posicaoUsuario']);
+
 });
 
 Route::prefix('compra') -> group(function (){
     Route::get('', [CompraController::class, 'listar']);
     Route::get('/{id}', [CompraController::class, 'buscarId']);
-    Route::post('/criar', [CompraController::class, 'criar']);
-    Route::put('/atualizar/{id}', [CompraController::class, 'atualizar']);
-    Route::delete('/deletar/{id}', [CompraController::class, 'deletar']);
-    Route::delete('/destroy/{id}', [CompraController::class, 'destroy']);
-    Route::post('/restore/{id}', [CompraController::class, 'restore']);
+    Route::middleware('auth:sanctum')->post('/criar', [CompraController::class, 'criar']);
+    Route::middleware('auth:sanctum')->put('/atualizar/{id}', [CompraController::class, 'atualizar']);
+    Route::middleware('auth:sanctum')->delete('/deletar/{id}', [CompraController::class, 'deletar']);
+    Route::middleware(['auth:sanctum', 'admin'])->delete('/destroy/{id}', [CompraController::class, 'destroy']);
+    Route::middleware(['auth:sanctum', 'admin'])->post('/restore/{id}', [CompraController::class, 'restore']);
 });
 
-// Route::prefix('login') -> group(function (){
-//     Route::get('', [LoginController::class, 'login']);
-// });
+Route::prefix('login') -> group(function (){
+    Route::post('', [LoginController::class, 'login']);
+    Route::middleware('auth:sanctum')->get('/verificarToken', [LoginController::class, 'verificarToken']);
+    Route::middleware('auth:sanctum')->post('/logout', [LoginController::class, 'logout']);
+});
